refactor: review반영
