+++ src/pages/mypage/genre.tsx
@@ -10,12 +10,9 @@ export default function Genre() {
     setEditMode((prev) => !prev);
   };
 
-  const getButtonLayoutClass = () => {
-    
-      return 'pc:grid pc:grid-cols-10 pc:gap-4 tablet:grid tablet:grid-cols-6 tablet:gap-4 mobile:grid mobile:grid-cols-3 mobile:gap-4';
-    
-  };
-
+  const getButtonLayoutClass = () => {      
+      return 'pc:grid pc:grid-cols-10 pc:gap-4 tablet:grid tablet:grid-cols-6 tablet:gap-4 mobile:grid mobile:grid-cols-3 mobile:gap-4';     
+  };  
   return (
     <div className='flex-col'>
       <div>선호장르 선택</div>
